requirements.txt
ffmpeg_python==0.2.0
numpy==1.24.1
omegaconf==2.3.0
openai_whisper==20231117
Pillow==10.0.1
Pillow==9.3.0
Pillow==10.2.0
playsound==1.2.2
playsound==1.3.0
Requests==2.31.0
sounddevice==0.4.6
soundfile==0.12.1
torch==2.1.2
torch==2.0.1
torch==2.1.1+cu121
torchaudio==2.1.2
torchaudio==2.0.2
torchaudio==2.1.1+cu121
